--- markku.py
@@ -1,540 +1,0 @@
-# coding: utf-8
-
-import json
-import logging
-import random
-import re
-from collections import Counter
-from os import environ
-from urllib.request import Request, urlopen
-
-from pymongo import ASCENDING, MongoClient
-from pymongo import errors as MongoErrors
-from telegram import InlineKeyboardButton, InlineKeyboardMarkup
-from telegram.ext import (BaseFilter, CommandHandler, Filters, MessageHandler, Updater, CallbackQueryHandler)
-
-# TODO: var -> jotkut vakiomuuttujat tähän
-
-def start(bot, update):
-    printlog(update, "start")
-
-    _, chat_id = get_ids(update) # Ignoraa user_id, tätä käytetään paljon
-    count_and_write(update, "commands")
-
-    bot.send_message(chat_id=chat_id, text="Woof woof")
-
-    
-# Lukee netistä valosensorin datan ja kertoo onko kerhohuoneella valot päällä
-def darkroom(bot, update):
-    printlog(update, "darkroom")
-
-    _, chat_id = get_ids(update)
-    count_and_write(update, "commands")
-    
-    with urlopen("https://ttkamerat.fi/darkroom/api/v1/sensors/latest") as url:
-        sensor_data = json.loads(url.read().decode())
-
-        value_light = 0
-        isDarkroomPopulated = False
-
-        # JSON härössä muodossa, sen takia teemme näin. Esimerkki:
-        #   {"entries": [{"value": 191, "sensor": "light1", "inserted": "2018-07-27T16:18:43.589Z"}]}
-        for sensor in sensor_data["entries"]:
-            if sensor["sensor"] == "light1":
-                value_light = sensor["value"]
-
-        if value_light > 100:
-            isDarkroomPopulated = True
-
-        if isDarkroomPopulated:
-            reply = "Joku on pimiöllä :O\n"
-        else:
-            reply = "Pimiö tyhjä :(\n"
-
-        bot.send_message(chat_id=chat_id, text=reply)
-
-
-def help(bot, update):
-    printlog(update, "help")
-
-    _, chat_id = get_ids(update)
-    count_and_write(update, "commands")
-
-    reply = "Komennot:\n" \
-            "/darkroom - Kertoo onko joku pimiöllä\n" \
-            "/stats - Chattikohtaiset statsit\n" \
-            "/toptenmsg - Chatin kovimmat viestittelijät\n"\
-            "/toptenkiitos - Chatin kovimmat kiitostelijat\n"\
-            "/noutaja - Postaa satunnaisen noutajakuvan\n"\
-            "/protip - Antaa ammatti valo kuvaus vinkin!\n"\
-            "/blacklist - Poista omat tietosi Markun tietokannasta ja estä uusien tallentaminen, lähetä privana Markulle\n"\
-            "/unblacklist - Salli omien tietojesi tallentaminen blacklist-komennon jälkeen, lähetä privana Markulle\n"\
-            "\n" \
-            "Botin koodit: @eltsu7 ja @kulmajaba\n" \
-            "Valosensorit ja siihen koodit: @anttimoi"
-
-    bot.send_message(chat_id=chat_id, text=reply)
-
-
-# Palauta updatesta user_id ja chat_id
-def get_ids(update):
-    # priva-chateissa chat id == user id
-    # TG API dokkareissa ei mainita kenttää from_user, mut se toimii joten ¯\_(ツ)_/¯
-    # oikee kenttä olis message.from.id mutta python vetää herneen nenään
-    return update.message.from_user.id, update.message.chat.id
-
-
-def count_and_write(update, var):
-    user_id, chat_id = get_ids(update)
-
-    # Älä laske blacklistattuja
-    if (blacklist_collection.find_one({ "user_id": user_id }) != None):
-        return
-
-    # TODO: siirrä setOnInsertin sisälle
-    # TODO: tsekkaa onko nimi Not Found, tsekkaa onko käyttäjänimi muuttunu järkeväks, jos on niin päivitä
-    username = "Not found"
-    if update.message.from_user.username is not None:
-        username = update.message.from_user.username
-
-    # countIncrementer kertoo $inc-operaattorille, mitä kasvatetaan ja kuinka paljon.
-    # Jos kenttää ei löydy, $inc luo sen samalla numeroarvolla
-    countIncrementer = {
-        "count.messages": 0,
-        "count.stickers": 0,
-        "count.photos": 0,
-        "count.gifs": 0,
-        "count.commands": 0,
-        "count.kiitos": 0
-    }
-    countIncrementer["count." + var] = 1
-    
-    # Update_one päivittää yhden dokumentin, eli yhden käyttäjän yhdessä chatissa.
-    # Boolean lopussa on upsert-parametri, = jos queryn mätsäävää dokumenttia ei löydy, se luodaan.
-    # SetOnInsert kertoo mitä muita kenttiä tehdään, jos luodaan uusi dokumentti
-    chats_collection.update_one(
-        { "chat_id": chat_id, "user_id": user_id },
-        { 
-            "$inc": countIncrementer,
-            "$setOnInsert": {
-                "chat_title": update.message.chat.title,
-                "username": username,
-            }
-        },
-        True
-    )
-
-    # Luo compound index (voidaan käyttää vain toisella tai molemmilla parametreilla)
-    chats_collection.create_index([
-        ("chat_id", ASCENDING),
-        ("user_id", ASCENDING)
-    ], unique=True)
-
-
-def toptenlist(chat_id, var):
-    cursor = chats_collection.aggregate([
-        { "$match": { "chat_id": chat_id }},
-        { "$project": { "_id": 0, "username": 1, "count": "$count." + var }},
-        { "$sort": { "count": -1 }},
-        { "$limit": 10 }
-    ])
-
-    topten_sorted = list(cursor)
-
-    text = ""
-    number = 1
-
-    for user in topten_sorted:
-        text += str(number) + ". " + user["username"] + ": " + str(user["count"]) + "\n"
-        number += 1
-
-    return text, len(topten_sorted)
-
-
-def topten_messages(bot, update):
-    printlog(update, "toptenmessages")
-
-    _, chat_id = get_ids(update)
-    count_and_write(update, "commands")
-
-    list, number = toptenlist(chat_id, "messages")
-
-    text = "Top " + str(number) + " viestittelijät:\n" + list
-
-    bot.send_message(chat_id=chat_id, text=text)
-
-
-def topten_kiitos(bot, update):
-    printlog(update, "toptenkiitos")
-
-    _, chat_id = get_ids(update)
-    count_and_write(update, "commands")
-
-    list, number = toptenlist(chat_id, "kiitos")
-
-    text = "Top " + str(number) + " kiitostelijat:\n" + list
-
-    bot.send_message(chat_id=chat_id, text=text)
-
-
-def noutaja(bot, update):
-    printlog(update, "noutaja")
-
-    _, chat_id = get_ids(update)
-    count_and_write(update, "commands")
-
-    url = "https://dog.ceo/api/breed/retriever/golden/images/random"
-
-    req = Request(url, headers={'User-Agent': 'Mozilla/5.0'})
-
-    with urlopen(req) as page:
-        retriever_data = json.loads(page.read().decode())
-
-        picture_link = retriever_data["message"]
-
-        bot.sendPhoto(chat_id=chat_id, photo=picture_link)
-
-
-def protip(bot, update):
-    printlog(update, "protip")
-
-    _, chat_id = get_ids(update)
-    count_and_write(update, "commands")
-
-    protip_index = random.randint(0, len(protip_list) - 1)
-
-    bot.send_message(chat_id=chat_id, text=protip_list[protip_index])
-
-
-def msg_text(bot, update):
-    printlog(update, "text")
-
-    _, chat_id = get_ids(update)
-    count_and_write(update, "messages")
-
-    message = update.message.text.lower()
-
-    parse_and_count(update)
-
-    lotto = random.randint(1, 201)
-
-    if "kiitos" in message:
-        count_and_write(update, "kiitos")
-
-        if lotto < 11:
-            update.message.reply_text("Kiitos")
-        elif lotto < 16:
-            sticker_index = random.randint(0, len(sticker_list) - 1)
-            bot.send_sticker(chat_id=chat_id, sticker=sticker_list[sticker_index])
-
-        elif lotto < 17:
-            update.message.reply_text("Ole hyvä")
-
-    elif "markku" in message and "istu" in message:
-        if lotto < 91:
-            bot.send_message(chat_id=chat_id, text="*istuu*")
-        else:
-            bot.send_message(chat_id=chat_id, text="*paskoo lattialle*")
-
-    elif "huono markku" in message:
-        bot.send_message(chat_id=chat_id, text="w00F")
-
-    elif "markku perkele" in message:
-        bot.send_message(chat_id=chat_id, text="woof?")
-
-    elif "filmi" in message and lotto < 11:
-        bot.send_message(chat_id=chat_id, text="Filmi best")
-
-
-def parse_and_count(update):
-    user_id, chat_id = get_ids(update)
-
-    # Älä laske blacklistattuja
-    if (blacklist_collection.find_one({ "user_id": user_id }) != None):
-        return
-
-    text = update.message.text.upper()
-
-    # usernameksi laitetaan 'Not found' jos sitä ei ole
-    username = "Not found"
-    if update.message.from_user.username is not None:
-        username = update.message.from_user.username
-
-    # chat titleksi laitetaan 'Private' jos sitä ei ole (priva chätti)
-    chat_title = "Private"
-    if update.message.chat.title is not None:
-        chat_title = update.message.chat.title
-    
-    # muuttaa kaikki paitsi kirjaimet ja numerot välilyönneiksi
-    parsed_text = re.sub('[^a-zA-Z0-9 öÖäÄ\n]', ' ', text)
-
-    # splittaa välilyöntien kohdalta
-    split_text = parsed_text.split(" ")
-
-    # poistaa listasta yhden ja nollan pituiset alkiot
-    split_text = [i for i in split_text if len(i) > 1]
-
-    # laskee listasta sanat ja tallentaa sen muotoon {"sana1": sanaMäärä1, "sana2": sanaMäärä2 ... }
-    wordCount = Counter(split_text)
-
-    # lisää joka sanan alkuun "words."
-    countIncrement = {}
-    for i in wordCount:
-        countIncrement["words." + i] = wordCount[i]
-
-    words_collection.update_one(
-        { "chat_id": chat_id, "user_id": user_id },
-        { 
-            "$inc": countIncrement,
-            "$setOnInsert": {
-                "chat_title": chat_title,
-                "username": username,
-            }
-        },
-        True
-    )
-
-
-def msg_sticker(bot, update):
-    printlog(update, "sticker")
-
-    count_and_write(update, "stickers")
-
-
-def msg_photo(bot, update):
-    printlog(update, "photo")
-
-    count_and_write(update, "photos")
-
-
-def msg_gif(bot, update):
-    printlog(update, "gif")
-
-    count_and_write(update, "gifs")
-
-
-def stats(bot, update):
-    printlog(update, "stats")
-
-    user_id, chat_id = get_ids(update)
-    # Älä poista tai muuta paikkaa ellet korjaa samalla, miten Markku käsittelee käyttäjän jota ei ole blacklistattu,
-    # mutta jolla ei ole vielä mitään statseja
-    count_and_write(update, "commands")
-
-    user = chats_collection.find_one({ "chat_id": chat_id, "user_id": user_id })
-
-    if (user == None):
-        update.message.reply_text("Markku ei seuraa sinua. Käytä komentoa /unblacklist , jos haluat seurannan käyttöön.\n" \
-                                  "Markku does not track you. Use the command /unblacklist to enable tracking.")
-
-    user_data = user["count"]
-
-    sticker_percent = "?"
-    kiitos_percent = "?"
-
-    if user_data["stickers"] + user_data["messages"] != 0:
-        sticker_percent = round(((user_data["stickers"]) / (user_data["stickers"] + user_data["messages"]) * 100), 2)
-
-    if user_data["messages"] != 0:
-        kiitos_percent = round(((user_data["kiitos"]) / (user_data["messages"]) * 100), 2)
-
-    msg = "@{}:\nMessages: {}".format(user["username"], user_data["messages"])
-    msg += "\nStickers: {} ({}%)".format(user_data["stickers"], sticker_percent)
-    msg += "\nKiitos: {} ({}%)".format(user_data["kiitos"], kiitos_percent)
-    msg += "\nPhotos: {}".format(user_data["photos"])
-
-    bot.send_message(chat_id=chat_id, text=msg)
-
-
-def status_new_members(bot, update):
-    printlog(update, "new member")
-
-    msg = camera_versus_text()
-
-    update.message.reply_text(msg)
-
-
-def camera_versus(bot, update):
-    printlog(update, "camera versus")
-
-    _, chat_id = get_ids(update)
-    count_and_write(update, "commands")
-
-    msg = camera_versus_text()
-    
-    bot.send_message(chat_id=chat_id, text=msg)
-
-
-def camera_versus_text():
-    # Painotettu lista random pickeihin
-
-    global camera_list
-    weighted_camera_list = camera_list["Common"] * 10 + camera_list["Medium"] * 3 + camera_list["Rare"]
-            
-    return "{} vai {}?".format(random.choice(weighted_camera_list), random.choice(weighted_camera_list))
-
-
-def blacklist(bot, update):
-    printlog(update, "blacklist")
-
-    user_id, _ = get_ids(update)
-
-    if (update.message.chat.type != "private"):
-        update.message.reply_text("Ole hyvä ja lähetä tämä pyyntö yksityisviestillä.\n" \
-                                  "Please send this request via private message.")
-        return
-    
-    if (blacklist_collection.find_one({ "user_id": user_id }) != None):
-        update.message.reply_text("Tietosi on jo poistettu, eikä sinua seurata.\n" \
-                                  "Your information is already deleted and you are not tracked.")
-        return
-    
-    keyboard = [[InlineKeyboardButton("Ei (No)", callback_data="false"),
-                 InlineKeyboardButton("Kyllä (Yes)", callback_data="true")]]
-
-    reply_markup = InlineKeyboardMarkup(keyboard)
-
-    update.message.reply_text("Haluatko, että kaikki henkilöivät tietosi poistetaan Markun tietokannasta," \
-                              "ja käyttäjäsi lisätään \"älä seuraa\"-listalle?\n\n" \
-                              "Do you want to delete all of your personifiable information from Markku's database" \
-                              "and add your user to the \"do not track\" list?",
-                              reply_markup=reply_markup)
-
-
-def blacklist_confirm(bot, update):
-    query = update.callback_query
-
-    username = query.from_user.username
-    user_id = query.from_user.id
-
-    print("Type: blacklist_confirm", "\nUsername: ", username)
-
-    if (query.data == "false"):
-        bot.edit_message_text(text="Käyttäjätietojen poistaminen peruttu.\n" \
-                                   "The deletion of user information has been cancelled.",
-                              chat_id=query.message.chat_id,
-                              message_id=query.message.message_id)
-        return
-    
-    bot.edit_message_text(text="Käyttäjätiedot poistettu, ja käyttäjää ei seurata jatkossa.\n" \
-                                "User information deleted, and the user will not be tracked.",
-                            chat_id=query.message.chat_id,
-                            message_id=query.message.message_id)
-
-    # Poista kaikki käyttäjän dokumentit
-    chats_collection.delete_many(
-        { "user_id": user_id }
-    )
-    words_collection.delete_many(
-        { "user_id": user_id }
-    )
-
-    try:
-        blacklist_collection.insert_one({
-            "user_id": user_id
-        })
-        blacklist_collection.create_index([
-            ("user_id", ASCENDING)
-        ], unique=True)
-    except MongoErrors.DuplicateKeyError:
-        print("User {} already blacklisted".format(username))
-
-
-def unblacklist(bot, update):
-    user_id, _ = get_ids(update)
-
-    if (update.message.chat.type != "private"):
-        update.message.reply_text("Ole hyvä ja lähetä tämä pyyntö yksityisviestillä\n" \
-                                  "Please send this request via private message")
-        return
-
-    deleteResult = blacklist_collection.delete_one(
-        { "user_id": user_id }
-    )
-
-    if (deleteResult.deleted_count != 0):
-        update.message.reply_text("Markku seuraa sinua taas.\n" \
-                                  "Markku is tracking you again.\n\n" \
-                                  "*sniff sniff* Woof!")
-    else:
-        update.message.reply_text("Et ole Markun \"älä seuraa\"-listalla.\n" \
-                                  "You are not on Markku's \"do not track\" list.")
-
-    
-def handlers(updater):
-    dp = updater.dispatcher
-
-    # Tässä alla oleville komennoille (esim darkroom) annetaan aina bot ja updater argumenteiksi
-    dp.add_handler(CommandHandler('start', start))
-    dp.add_handler(CommandHandler('darkroom', darkroom))
-    dp.add_handler(CommandHandler('stats', stats))
-    dp.add_handler(CommandHandler('help', help))
-    dp.add_handler(CommandHandler('noutaja', noutaja))
-    dp.add_handler(CommandHandler('toptenmsg', topten_messages))
-    dp.add_handler(CommandHandler('toptenkiitos', topten_kiitos))
-    dp.add_handler(CommandHandler('protip', protip))
-    dp.add_handler(CommandHandler('kysymys', camera_versus))
-    dp.add_handler(CommandHandler('blacklist', blacklist))
-    dp.add_handler(CommandHandler('unblacklist', unblacklist))
-    dp.add_handler(CallbackQueryHandler(blacklist_confirm))
-    dp.add_handler(MessageHandler(Filters.sticker, msg_sticker))
-    dp.add_handler(MessageHandler(Filters.text, msg_text))
-    dp.add_handler(MessageHandler(Filters.photo, msg_photo))
-    dp.add_handler(MessageHandler(Filters.document, msg_gif))
-    dp.add_handler(MessageHandler(Filters.status_update.new_chat_members, status_new_members))
-
-
-def printlog(update, msg_type):
-    username = update.message.from_user.username
-    content = ""
-
-    print("Type: ", msg_type, "\nUsername: ", username)
-
-    if msg_type == "sticker":
-        content = update.message.sticker.file_id
-    elif msg_type == "text":
-        content = update.message.text
-
-    if content != "":
-        print("Content: ", content)
-
-    print()
-
-
-# Lue JSON-tiedosto
-def file_read(filename):
-    try:
-        with open(filename, 'r', encoding='utf-8') as file:
-            saved_data = json.load(file)
-        file.close()
-        return saved_data
-    except FileNotFoundError:
-        print("Oh dog file not found")
-        exit(1)
-
-
-def main():
-    updater = Updater(token=tg_token)
-    handlers(updater)
-
-    updater.start_polling()
-
-masterlist = file_read("masterlist.json")
-sticker_list = masterlist["Stickers"]
-protip_list = masterlist["Tips"]
-camera_list = masterlist["Cameras"]
-
-tg_token = environ["TG_TOKEN"]
-db_name = environ["DB_NAME"]
-chats_coll_name = environ["CHATS_COLL_NAME"]
-words_coll_name = environ["WORDS_COLL_NAME"]
-blacklist_coll_name = environ["BLACKLIST_COLL_NAME"]
-
-# TODO: failaa jos ei saada yhteyttä
-db_client = MongoClient("mongodb://mongo:27017", serverSelectionTimeoutMS=1000)
-db = db_client[db_name]
-chats_collection = db[chats_coll_name]
-words_collection = db[words_coll_name]
-blacklist_collection = db[blacklist_coll_name]
-
-main()
