+++ src/core/file_read.py
@@ -1,0 +1,12 @@
+import json
+
+# Lue JSON-tiedosto
+def file_read(filename):
+    try:
+        with open(filename, 'r', encoding='utf-8') as file:
+            saved_data = json.load(file)
+        file.close()
+        return saved_data
+    except FileNotFoundError:
+        print("Oh dog file not found")
+        exit(1)
