+++ src/message_handlers/msg_photo.py
@@ -1,0 +1,7 @@
+from core.printlog import printlog
+from core.count_and_write import count_and_write
+
+def msg_photo(bot, update):
+    printlog(update, "photo")
+
+    count_and_write(update, "photos")
