+++ src/message_handlers/status_new_members.py
@@ -1,0 +1,9 @@
+from core.printlog import printlog
+from core.camera_versus_text import camera_versus_text
+
+def status_new_members(bot, update):
+    printlog(update, "new member")
+
+    msg = camera_versus_text()
+
+    update.message.reply_text(msg)
